Feature/refactoring (#6)
* move original app to new dir

* move uivicon file

* fix error

* remove component

* add root router

* add pipeline router

* add feature upload file to api
