--- git_up/main.py
@@ -1,52 +1,0 @@
-from dataclasses import asdict
-from typing import Optional
-
-import uvicorn
-from fastapi import FastAPI, UploadFile, File 
-import shutil
-from app.database.conn import db
-from app.common.config import conf
-from app.routes import index, auth
-aa
-
-
-def create_app():
-    """
-    앱 함수 실행
-    :return:
-    """
-    c = conf()
-    app = FastAPI()
-    conf_dict = asdict(c)
-    db.init_app(app, **conf_dict)
-    # 데이터 베이스 이니셜라이즈
-
-    # 레디스 이니셜라이즈
-
-    # 미들웨어 정의
-
-    # 라우터 정의
-    app.include_router(index.router)
-    app.include_router(auth.router, tags=["Authentication"], prefix="/auth")
-    return app
-
-
-app = create_app()
-
-if __name__ == "__main__":
-    uvicorn.run("main:app", host="0.0.0.0", port=8080, reload=True)
-
-@app.post("/")
-async def root(file: UploadFile = File(...)):
-    with open(f'{file.filename}', "wb") as buffer:
-        shutil.copyfileobj(file.file, buffer)
-    return {"file_name": file.filename}
-
-@app.post("/img")
-async def upload_image(files: List[UploadFile] = File(...)):
-    for img in files:
-        with open(f'{img.filename}', "wb") as buffer:
-            shutil.copyfileobj(img.file, buffer)
-    
-    return {"file_name": "good"}
-
